+++ src/edge-core-index.js
@@ -26,29 +26,25 @@ export {
   SameCurrencyError,
   UsernameError
 } from './error.js'
-export {
-  makeContext,
-  makeEdgeContext,
-  makeFakeContexts
-} from './makeContext.js'
+export { makeEdgeContext, makeFakeContexts } from './makeContext.js'
 export { destroyAllContexts } from './modules/root.js'
 
 // io types -----------------------------------------------------------
 
-export interface DiskletFile {
-  delete(): Promise<void>;
-  getData(): Promise<Uint8Array>;
-  getText(): Promise<string>;
-  setData(data: Array<number> | Uint8Array): Promise<void>;
-  setText(text: string): Promise<void>;
+export type DiskletFile = {
+  delete(): Promise<mixed>,
+  getData(): Promise<Uint8Array>,
+  getText(): Promise<string>,
+  setData(data: Array<number> | Uint8Array): Promise<mixed>,
+  setText(text: string): Promise<mixed>
 }
 
-export interface DiskletFolder {
-  delete(): Promise<void>;
-  file(name: string): DiskletFile;
-  folder(name: string): DiskletFolder;
-  listFiles(): Promise<Array<string>>;
-  listFolders(): Promise<Array<string>>;
+export type DiskletFolder = {
+  delete(): Promise<mixed>,
+  file(name: string): DiskletFile,
+  folder(name: string): DiskletFolder,
+  listFiles(): Promise<Array<string>>,
+  listFolders(): Promise<Array<string>>
 }
 
 // Node.js randomBytes function:
@@ -98,29 +94,6 @@ export type EdgePbkdf2 = {
 }
 
 /**
- * Access to platform-specific resources, with many optional fields.
- * The core will emulate/adapt whatever is missing.
- */
-export type EdgeRawIo = {
-  // Crypto:
-  +random: EdgeRandomFunction, // Non-optional & security-critical
-  +scrypt?: EdgeScryptFunction,
-  +secp256k1?: EdgeSecp256k1,
-  +pbkdf2?: EdgePbkdf2,
-
-  // Local io:
-  +console?: EdgeConsole,
-  +folder?: DiskletFolder,
-  +localStorage?: Storage,
-
-  // Networking:
-  +fetch: typeof fetch,
-  +Socket?: typeof net$Socket,
-  +TLSSocket?: typeof tls$TLSSocket,
-  +WebSocket: typeof WebSocket
-}
-
-/**
  * Access to platform-specific resources.
  * The core never talks to the outside world on its own,
  * but always goes through this object.
@@ -165,26 +138,22 @@ export type EdgeContextOptions = {
   appId?: string,
   authServer?: string,
   callbacks?: EdgeContextCallbacks,
-  io?: EdgeRawIo,
   path?: string, // Only used on node.js
   plugins?: Array<EdgeCorePluginFactory>,
   shapeshiftKey?: string,
 
   // Used by the fake context:
   localFakeUser?: boolean,
-
-  // Deprecated. Use appId instead:
-  accountType?: string
+  offline?: boolean
 }
 
 export type EdgeContext = {
-  appId: string,
-  io: EdgeIo,
+  +appId: string,
 
   // Local user management:
   fixUsername(username: string): string,
   listUsernames(): Promise<Array<string>>,
-  deleteLocalAccount(username: string): Promise<void>,
+  deleteLocalAccount(username: string): Promise<mixed>,
 
   // Account creation:
   usernameAvailable(username: string): Promise<boolean>,
@@ -238,9 +207,6 @@ export type EdgeContext = {
   // OTP stuff:
   requestOtpReset(username: string, otpResetToken: string): Promise<Date>,
   fetchLoginMessages(): Promise<EdgeLoginMessages>,
-
-  // Misc. stuff:
-  getCurrencyPlugins(): Promise<Array<EdgeCurrencyPlugin>>,
 
   // Shapeshift:
   getExchangeSwapRate(
@@ -361,8 +327,11 @@ export type EdgeCreateCurrencyWalletOptions = {
   keys?: {}
 }
 
-export interface EdgeCurrencyTools {
-  +currencyInfo: EdgeCurrencyInfo;
+export type EdgeCurrencyTools = {
+  +currencyInfo: EdgeCurrencyInfo,
+  +pluginSettings: Object,
+
+  changePluginSettings(settings: Object): Promise<mixed>
 }
 
 export type EdgeCurrencyToolsMap = {
@@ -395,14 +364,14 @@ export type EdgeAccount = {
 
   // What login method was used?
   +edgeLogin: boolean,
-  keyLogin: boolean,
-  newAccount: boolean,
-  passwordLogin: boolean,
-  pinLogin: boolean,
-  recoveryLogin: boolean,
+  +keyLogin: boolean,
+  +newAccount: boolean,
+  +passwordLogin: boolean,
+  +pinLogin: boolean,
+  +recoveryLogin: boolean,
 
   // Change or create credentials:
-  changePassword(password: string): Promise<void>,
+  changePassword(password: string): Promise<mixed>,
   changePin(opts: {
     pin?: string, // We keep the existing PIN if unspecified
     enableLogin?: boolean // We default to true if unspecified
@@ -417,31 +386,31 @@ export type EdgeAccount = {
   checkPin(pin: string): Promise<boolean>,
 
   // Remove credentials:
-  deletePassword(): Promise<void>,
-  deletePin(): Promise<void>,
-  deleteRecovery(): Promise<void>,
+  deletePassword(): Promise<mixed>,
+  deletePin(): Promise<mixed>,
+  deleteRecovery(): Promise<mixed>,
 
   // OTP:
   +otpKey: string | void, // OTP is enabled if this exists
   +otpResetDate: string | void, // A reset is requested if this exists
-  cancelOtpReset(): Promise<void>,
-  disableOtp(): Promise<void>,
-  enableOtp(timeout?: number): Promise<void>,
+  cancelOtpReset(): Promise<mixed>,
+  disableOtp(): Promise<mixed>,
+  enableOtp(timeout?: number): Promise<mixed>,
 
   // Edge login approval:
   fetchLobby(lobbyId: string): Promise<EdgeLobby>,
 
   // Login management:
-  logout(): Promise<void>,
+  logout(): Promise<mixed>,
 
   // Master wallet list:
   +allKeys: Array<EdgeWalletInfoFull>,
-  changeWalletStates(walletStates: EdgeWalletStates): Promise<void>,
+  changeWalletStates(walletStates: EdgeWalletStates): Promise<mixed>,
   createWallet(type: string, keys: any): Promise<string>,
   getFirstWalletInfo(type: string): ?EdgeWalletInfo,
   getWalletInfo(id: string): ?EdgeWalletInfo,
   listWalletIds(): Array<string>,
-  listSplittableWalletTypes(walletId: string): Array<string>,
+  listSplittableWalletTypes(walletId: string): Promise<Array<string>>,
   splitWalletInfo(walletId: string, newWalletType: string): Promise<string>,
 
   // Currency wallets:
@@ -457,16 +426,16 @@ export type EdgeAccount = {
 // edge login types ---------------------------------------------------
 
 export type EdgeLobby = {
-  loginRequest?: EdgeLoginRequest
-  // walletRequest?: EdgeWalletRequest
+  +loginRequest: EdgeLoginRequest | void
+  // walletRequest: EdgeWalletRequest | void
 }
 
 export type EdgeLoginRequest = {
-  appId: string,
-  approve(): Promise<void>,
+  +appId: string,
+  approve(): Promise<mixed>,
 
-  displayName: string,
-  displayImageUrl?: string
+  +displayName: string,
+  +displayImageUrl: string | void
 }
 
 // currency wallet types ----------------------------------------------
@@ -477,6 +446,8 @@ export type EdgeTokenInfo = {
   contractAddress: string,
   multiplier: string
 }
+
+export type EdgeBalances = { [currencyCode: string]: string }
 
 export type EdgeGetTransactionsOptions = {
   currencyCode?: string,
@@ -503,66 +474,68 @@ export type EdgeCurrencyWallet = {
   +id: string,
   +keys: any,
   +type: string,
+  +displayPrivateSeed: string | null,
+  +displayPublicSeed: string | null,
 
   // Data store:
   +folder: DiskletFolder,
   +localFolder: DiskletFolder,
-  sync(): Promise<void>,
+  sync(): Promise<mixed>,
 
   // Wallet name:
   +name: string | null,
-  renameWallet(name: string): Promise<void>,
+  renameWallet(name: string): Promise<mixed>,
 
   // Fiat currency option:
   +fiatCurrencyCode: string,
-  setFiatCurrencyCode(fiatCurrencyCode: string): Promise<void>,
+  setFiatCurrencyCode(fiatCurrencyCode: string): Promise<mixed>,
 
   // Currency info:
   +currencyInfo: EdgeCurrencyInfo,
 
+  // Chain state:
+  +balances: EdgeBalances,
+  +blockHeight: number,
+
   // Running state:
-  startEngine(): Promise<void>,
-  stopEngine(): Promise<void>,
+  startEngine(): Promise<mixed>,
+  stopEngine(): Promise<mixed>,
 
   // Token management:
-  enableTokens(tokens: Array<string>): Promise<void>,
-  disableTokens(tokens: Array<string>): Promise<void>,
+  enableTokens(tokens: Array<string>): Promise<mixed>,
+  disableTokens(tokens: Array<string>): Promise<mixed>,
   getEnabledTokens(): Promise<Array<string>>,
-  addCustomToken(token: EdgeTokenInfo): Promise<void>,
+  addCustomToken(token: EdgeTokenInfo): Promise<mixed>,
 
   // Transactions:
-  getBalance(opts?: EdgeCurrencyCodeOptions): string,
-  getBlockHeight(): number,
-  getNumTransactions(opts?: EdgeCurrencyCodeOptions): number,
+  getNumTransactions(opts?: EdgeCurrencyCodeOptions): Promise<number>,
   getTransactions(
     options?: EdgeGetTransactionsOptions
   ): Promise<Array<EdgeTransaction>>,
   getReceiveAddress(
     opts?: EdgeCurrencyCodeOptions
   ): Promise<EdgeReceiveAddress>,
-  saveReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<void>,
-  lockReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<void>,
-  makeAddressQrCode(address: EdgeReceiveAddress): string,
-  makeAddressUri(address: EdgeReceiveAddress): string,
+  saveReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<mixed>,
+  lockReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<mixed>,
   makeSpend(spendInfo: EdgeSpendInfo): Promise<EdgeTransaction>,
   signTx(tx: EdgeTransaction): Promise<EdgeTransaction>,
   broadcastTx(tx: EdgeTransaction): Promise<EdgeTransaction>,
-  saveTx(tx: EdgeTransaction): Promise<void>,
+  saveTx(tx: EdgeTransaction): Promise<mixed>,
   sweepPrivateKeys(edgeSpendInfo: EdgeSpendInfo): Promise<EdgeTransaction>,
   saveTxMetadata(
     txid: string,
     currencyCode: string,
     metadata: EdgeMetadata
-  ): Promise<void>,
+  ): Promise<mixed>,
   getMaxSpendable(spendInfo: EdgeSpendInfo): Promise<string>,
   getQuote(spendInfo: EdgeSpendInfo): Promise<EdgeCoinExchangeQuote>,
-  getPaymentProtocolInfo?: (
+  getPaymentProtocolInfo(
     paymentProtocolUrl: string
-  ) => Promise<EdgePaymentProtocolInfo>,
+  ): Promise<EdgePaymentProtocolInfo>,
 
   // Wallet management:
-  resyncBlockchain(): Promise<void>,
-  dumpData(): EdgeDataDump,
+  resyncBlockchain(): Promise<mixed>,
+  dumpData(): Promise<EdgeDataDump>,
   getDisplayPrivateSeed(): string | null,
   getDisplayPublicSeed(): string | null,
 
@@ -571,8 +544,12 @@ export type EdgeCurrencyWallet = {
   exportTransactionsToCSV(opts: EdgeGetTransactionsOptions): Promise<string>,
 
   // URI handling:
-  parseUri(uri: string): EdgeParsedUri,
-  encodeUri(obj: EdgeEncodeUri): string
+  parseUri(uri: string): Promise<EdgeParsedUri>,
+  encodeUri(obj: EdgeEncodeUri): Promise<string>,
+
+  // Deprecated API's:
+  getBalance(opts?: EdgeCurrencyCodeOptions): string,
+  getBlockHeight(): number
 }
 
 export type EdgeMetadata = {
@@ -743,14 +720,13 @@ export type EdgeCurrencyEngineOptions = {
 }
 
 export type EdgeCurrencyEngine = {
-  updateSettings(settings: any): void,
-  startEngine(): Promise<void>,
-  killEngine(): Promise<void>,
+  startEngine(): Promise<mixed>,
+  killEngine(): Promise<mixed>,
   getBlockHeight(): number,
-  enableTokens(tokens: Array<string>): Promise<void>,
-  disableTokens(tokens: Array<string>): Promise<void>,
+  enableTokens(tokens: Array<string>): Promise<mixed>,
+  disableTokens(tokens: Array<string>): Promise<mixed>,
   getEnabledTokens(): Promise<Array<string>>,
-  addCustomToken(token: EdgeTokenInfo): Promise<void>,
+  addCustomToken(token: EdgeTokenInfo): Promise<mixed>,
   getTokenStatus(token: string): boolean,
   getBalance(options: EdgeCurrencyCodeOptions): string,
   getNumTransactions(options: EdgeCurrencyCodeOptions): number,
@@ -767,8 +743,8 @@ export type EdgeCurrencyEngine = {
   +sweepPrivateKeys?: (abcSpendInfo: EdgeSpendInfo) => Promise<EdgeTransaction>,
   signTx(abcTransaction: EdgeTransaction): Promise<EdgeTransaction>,
   broadcastTx(abcTransaction: EdgeTransaction): Promise<EdgeTransaction>,
-  saveTx(abcTransaction: EdgeTransaction): Promise<void>,
-  resyncBlockchain(): Promise<void>,
+  saveTx(abcTransaction: EdgeTransaction): Promise<mixed>,
+  resyncBlockchain(): Promise<mixed>,
   dumpData(): EdgeDataDump,
   +getPaymentProtocolInfo?: (
     paymentProtocolUrl: string
@@ -800,11 +776,12 @@ export type EdgeCurrencyPlugin = {
   ): Promise<EdgeCurrencyEngine>,
   parseUri(uri: string): EdgeParsedUri,
   encodeUri(obj: EdgeEncodeUri): string,
-  +getSplittableTypes?: (walletInfo: EdgeWalletInfo) => Array<string>
+  +getSplittableTypes?: (walletInfo: EdgeWalletInfo) => Array<string>,
+  +changeSettings?: (settings: Object) => Promise<mixed>
 }
 
 export type EdgeCurrencyPluginFactory = {
-  pluginType: 'currency',
+  +pluginType: 'currency',
   +pluginName: string,
   makePlugin(opts: EdgeCorePluginOptions): Promise<EdgeCurrencyPlugin>
 }
@@ -823,7 +800,7 @@ export type EdgeExchangePair = {
 }
 
 export type EdgeExchangePlugin = {
-  exchangeInfo: { exchangeName: string },
+  +exchangeInfo: { exchangeName: string },
 
   fetchExchangeRates(
     pairHints: Array<EdgeExchangePairHint>
@@ -831,7 +808,7 @@ export type EdgeExchangePlugin = {
 }
 
 export type EdgeExchangePluginFactory = {
-  pluginType: 'exchange',
+  +pluginType: 'exchange',
   makePlugin(opts: EdgeCorePluginOptions): Promise<EdgeExchangePlugin>
 }
 
@@ -842,7 +819,6 @@ export type {
   EdgeScryptFunction as AbcScryptFunction,
   EdgeSecp256k1 as AbcSecp256k1,
   EdgePbkdf2 as AbcPbkdf2,
-  EdgeRawIo as AbcRawIo,
   EdgeIo as AbcIo,
   EdgeCorePluginFactory as AbcCorePluginFactory,
   EdgeCorePluginOptions as AbcCorePluginOptions,