+++ src/internal.js
@@ -4,9 +4,6 @@
 // which makes use of some undocumented core features.
 // In the future we hope to minimize / reduce this
 
-export { authRequest } from './modules/login/authServer.js'
-export { hashUsername } from './modules/login/loginStore.js'
-export { makeLobby } from './modules/login/lobby.js'
 export { hmacSha256 } from './util/crypto/crypto.js'
 export * from './util/decorators.js'
 export * from './util/encoding.js'