+++ src/modules/context/context-api-pixie.js
@@ -12,7 +12,6 @@ import type {
 import { wrapObject } from '../../util/api.js'
 import { base58 } from '../../util/encoding.js'
 import { makeAccount } from '../account/account-api.js'
-import { waitForCurrencyPlugins } from '../currency/currency-selectors.js'
 import { makeShapeshiftApi } from '../exchange/shapeshift.js'
 import { createLogin, usernameAvailable } from '../login/create.js'
 import { requestEdgeLogin } from '../login/edge.js'
@@ -27,6 +26,7 @@ import {
   loginRecovery2
 } from '../login/recovery2.js'
 import type { ApiInput } from '../root.js'
+import { EdgeInternalStuff } from './internal-api.js'
 
 export const contextApiPixie = (ai: ApiInput) => () => {
   ai.onOutput(makeContextApi(ai))
@@ -36,15 +36,15 @@ export const contextApiPixie = (ai: ApiInput) => () => {
 function makeContextApi (ai: ApiInput) {
   const appId = ai.props.state.login.appId
   const { loginStore } = ai.props
+  const internalApi = new EdgeInternalStuff(ai)
 
   const shapeshiftApi = makeShapeshiftApi(ai)
 
   const rawContext: EdgeContext = {
-    io: (ai.props.io: any),
     appId,
 
-    getCurrencyPlugins () {
-      return waitForCurrencyPlugins(ai)
+    get _internalEdgeStuff () {
+      return internalApi
     },
 
     '@fixUsername': { sync: true },
@@ -56,7 +56,7 @@ function makeContextApi (ai: ApiInput) {
       return loginStore.listUsernames()
     },
 
-    deleteLocalAccount (username: string): Promise<void> {
+    deleteLocalAccount (username: string): Promise<mixed> {
       return loginStore.remove(username)
     },
 
@@ -221,11 +221,6 @@ function makeContextApi (ai: ApiInput) {
 
   // Wrap the context with logging:
   const out = wrapObject('Context', rawContext)
-  out.usernameList = out.listUsernames
-  out.removeUsername = out.deleteLocalAccount
-
-  // Used for the edge-login unit tests:
-  out.internalUnitTestingHack = () => ai
 
   return out
 }