+++ src/modules/currency/currency-reducer.js
@@ -8,11 +8,14 @@ import type { RootState } from '../root-reducer.js'
 import type { CurrencyWalletState } from './wallet/currency-wallet-reducer.js'
 import currencyWalletReducer from './wallet/currency-wallet-reducer.js'
 
+export type PluginSettings = { [pluginName: string]: Object }
+
 export interface CurrencyState {
   currencyWalletIds: Array<string>;
   customTokens: Array<EdgeTokenInfo>;
   infos: Array<EdgeCurrencyInfo>;
   pluginsError: Error | null;
+  settings: PluginSettings;
   wallets: { [walletId: string]: CurrencyWalletState };
 }
 
@@ -52,6 +55,20 @@ export default buildReducer({
 
   pluginsError (state = null, action: RootAction) {
     return action.type === 'CURRENCY_PLUGINS_FAILED' ? action.payload : state
+  },
+
+  settings (state: PluginSettings = {}, action: RootAction): PluginSettings {
+    switch (action.type) {
+      case 'CHANGED_CURRENCY_PLUGIN_SETTING':
+        const { pluginName, settings } = action.payload
+        const out = { ...state }
+        out[pluginName] = settings
+        return out
+
+      case 'NEW_CURRENCY_PLUGIN_SETTINGS':
+        return action.payload
+    }
+    return state
   },
 
   wallets: mapReducer(